FlashLoanMaster/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "bootstrap-flask>=2.4.2",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "web3==6.11.3",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.40",
    "bitcoinlib>=0.7.3",
    "reportlab>=4.3.1",
    "redis>=5.2.1",
    "prometheus-client>=0.21.1",
    "python-telegram-bot>=22.0",
]
